Wait for gnocchi service before configuring ceilometer
Currently the ceilometer services are configured and
ceilometer dbsync is run as an action.
A new interface gnocchi_service is introduced that
updates relation data when gnocchi service is up.
Add new relation gnocchi_service in metadata.yaml
as a mandatory relation.
Update dbsync commands so that they are run as part
of configure_charm.
Remove the action ceilometer-upgrade that does the
dbsync as it is no more required.

Change-Id: Iba72250f3d31ef9dafa5b3bd16abefd43a74366a